+++ src/main/java/controller/BlackjackController.java
@@ -4,8 +4,8 @@ import domain.RandomValueGenerator;
 import domain.card.Card;
 import domain.card.Deck;
 import domain.card.Hand;
-import domain.hitstategy.CasinoDealerHitStrategy;
-import domain.hitstategy.UntilBurstHitStrategy;
+import domain.hitStrategy.CasinoDealerHitStrategy;
+import domain.hitStrategy.UntilBurstHitStrategy;
 import domain.participants.Dealer;
 import domain.participants.Player;
 import domain.score.Result;
