+++ src/main/java/domain/hitStrategy/HitStrategy.java
@@ -1,4 +1,4 @@
-package domain.hitstategy;
+package domain.hitStrategy;
 
 import domain.score.Score;
