+++ src/main/java/domain/hitStrategy/UntilBurstHitStrategy.java
@@ -1,4 +1,4 @@
-package domain.hitstategy;
+package domain.hitStrategy;
 
 import domain.score.Score;
