+++ src/main/java/domain/participants/Dealer.java
@@ -1,7 +1,7 @@
 package domain.participants;
 
 import domain.card.Hand;
-import domain.hitstategy.HitStrategy;
+import domain.hitStrategy.HitStrategy;
 import domain.score.Score;
 import java.util.List;
